Add end-to-end and unit tests for AIChat functionality; remove obsolete tests and configurations
