+++ Backend/Dotnet.Test/Component/Service/Unit/Applications/AIChat/AIChatServiceHandlerTests.cs
@@ -6,8 +6,6 @@ using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
 using Xunit;
-using Moq;
-using Moq.Protected;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
