--- Backend/Dotnet.Test/UnitTest1.cs
@@ -1,10 +1,0 @@
-﻿namespace Dotnet.Test;
-
-public class DotnetProjectTests
-{
-    [Fact]
-    public void Test1()
-    {
-
-    }
-}
