--- Backend/Dotnet/appsettings.Development.json
@@ -1,37 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-
-  "AIChat": {
-    "E2E": {
-      "Provider": "Mock"
-    },
-    "ProviderInfo": {
-      "Mock": {
-        "BaseUrl": "local",
-        "ApiKey": "none",
-        "Model": "MockModelV1",
-        "MaxTokens": 100,
-        "TimeoutSeconds": 1
-      },
-      "OpenRouter": {
-        "BaseUrl": "https://openrouter.ai/api/",
-        "ApiKey": "",
-        "Model": "mistralai/mistral-7b-instruct:free",
-        "MaxTokens": 1000,
-        "TimeoutSeconds": 120
-      },
-      "OpenAi": {
-        "BaseUrl": "https://api.openai.com/",
-        "ApiKey": "",
-        "Model": "gpt-4o-mini",
-        "MaxTokens": 512,
-        "TimeoutSeconds": 60
-      }
-    }
-  }
-}
